forward pass of ai2v
